+++ tests/exec/imple_statedb_test.go
@@ -31,7 +31,7 @@ func TestStateDBV2GetNonexistBalance(t *testing.T) {
 	// fmt.Println("\n" + euCommon.FormatTransitions(transitions))
 	committer := stgcomm.NewStateCommitter(db, nil)
 	committer.Import(transitions)
-	committer.Precommit([]uint32{1})
+	committer.Precommit([]uint64{1})
 	committer.Commit(20)
 
 	committer = stgcomm.NewStateCommitter(db, nil)
@@ -62,7 +62,7 @@ func TestStateDBV2GetNonexistCode(t *testing.T) {
 
 	committer := stgcomm.NewStateCommitter(db, nil)
 	committer.Import(transitions)
-	committer.Precommit([]uint32{1})
+	committer.Precommit([]uint64{1})
 	committer.Commit(20)
 
 	committer = stgcomm.NewStateCommitter(db, nil)
@@ -93,7 +93,7 @@ func TestStateDBV2GetNonexistStorageState(t *testing.T) {
 	// fmt.Println("\n" + euCommon.FormatTransitions(transitions))
 	committer := stgcomm.NewStateCommitter(db, nil)
 	committer.Import(transitions)
-	committer.Precommit([]uint32{1})
+	committer.Precommit([]uint64{1})
 	committer.Commit(20)
 
 	committer = stgcomm.NewStateCommitter(db, nil)
@@ -124,7 +124,7 @@ func TestStateDBV2GetNonexistStorageState(t *testing.T) {
 // 	_, transitions := api.WriteCache().(*tempcache.WriteCache).ExportAll()
 // 	// fmt.Println("\n" + euCommon.FormatTransitions(transitions))
 // 	committer.Import(transitions)
-// 	committer.Precommit([]uint32{1})
+// 	committer.Precommit([]uint64{1})
 // 	committer.Commit(20)
 
 // 	committer = stgcomm.NewStateCommitter(db)
