+++ tests/storage/balance_nonce_test.go
@@ -76,7 +76,7 @@ func TestSimpleBalance(t *testing.T) {
 
 	committer := stgcommitter.NewStateCommitter(store, sstore.GetWriters())
 	committer.Import(out)
-	committer.Precommit([]uint32{stgcommcommon.SYSTEM, 0, 1})
+	committer.Precommit([]uint64{stgcommcommon.SYSTEM, 0, 1})
 	committer.Commit(10)
 	// Read alice's balance again
 
@@ -230,7 +230,7 @@ func TestNonce(t *testing.T) {
 	committer := stgcommitter.NewStateCommitter(store, sstore.GetWriters())
 	committer.Import(trans)
 
-	committer.Precommit([]uint32{0})
+	committer.Precommit([]uint64{0})
 	committer.Commit(10)
 	nonce, _, _ = writeCache.Read(0, "blcc://eth1.0/account/"+alice+"/nonce", new(commutative.Uint64))
 	v = nonce.(uint64)
@@ -301,7 +301,7 @@ func TestMultipleNonces(t *testing.T) {
 	}
 
 	// committer = stgcommitter.NewStateCommitter(store, sstore.GetWriters())
-	committer.Precommit([]uint32{0, stgcommcommon.SYSTEM})
+	committer.Precommit([]uint64{0, stgcommcommon.SYSTEM})
 	committer.Commit(10)
 
 	nonce, _, _ = writeCache.Read(0, "blcc://eth1.0/account/"+bob+"/nonce", new(commutative.Uint64))
@@ -329,7 +329,7 @@ func TestUint64Delta(t *testing.T) {
 	acctTrans := univalue.Univalues(slice.Clone(writeCache.Export(univalue.Sorter))).To(univalue.IPTransition{})
 
 	committer := stgcommitter.NewStateCommitter(store, sstore.GetWriters())
-	committer.Import(acctTrans).Precommit([]uint32{stgcommcommon.SYSTEM})
+	committer.Import(acctTrans).Precommit([]uint64{stgcommcommon.SYSTEM})
 	committer.Commit(stgcommcommon.SYSTEM)
 
 	deltav1 := commutative.NewUint64Delta(11)
@@ -353,7 +353,7 @@ func TestUint64Delta(t *testing.T) {
 	committer = stgcommitter.NewStateCommitter(store, sstore.GetWriters())
 	committer.Import(acctTrans0)
 	committer.Import(acctTrans1)
-	committer.Precommit([]uint32{1, 2})
+	committer.Precommit([]uint64{1, 2})
 	committer.Commit(10)
 
 }
